requirements.txt
requests
python-dotenv
openai
flask
gunicorn
pypdf
python-docx
